cambio de estrategia, ahora en memory full la memoria espera a que se le envíe el insert de nuevo.
hay que revisar los frees y el describe y el timestamp en funcionInsert

+++ Memoria/src/estructuras/paginas.h
@@ -39,7 +39,7 @@ typedef struct reg_tablaDePaginas {
     uint32_t numeroPagina;
     registro_tad* punteroAPagina;
     bool flagModificado;
-    uint32_t ultimoAcceso;
+    uint64_t ultimoAcceso;
 }reg_tablaDePaginas;
 
 // Estructura de la Tabla de Páginas dentro de un Segmento en una Sub-Memoria;
